requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==2.1.9
av==9.1.1
bitarray==2.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
git+https://github.com/openai/CLIP.git
colorama==0.4.4
colorlog==5.0.1
configparser==5.1.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.6
decorator==5.0.9
docker-pycreds==0.4.0
filelock==3.0.12
fire==0.4.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
huggingface-hub==0.5.1
HyperPyYAML==1.0.0
idna==3.2
importlib-resources==5.2.2
joblib==1.0.1
julius==0.2.6
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.3
numba==0.53.1
numpy==1.21.1
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.0
pandas==1.3.3
pathtools==0.1.2
pesq==0.0.3
Pillow==9.1.1
pip==21.2.4
pooch==1.4.0
portalocker==2.3.2
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pypesq==1.2.4
pystoi==0.3.3
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
sacrebleu==2.0.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
sentencepiece==0.1.96
sentry-sdk==1.5.0
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
sk-video==1.1.10
sklearn==0.0
smmap==5.0.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
sox==1.4.1
speechbrain==0.5.9
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
threadpoolctl==2.2.0
timm==0.4.12
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio
tqdm==4.62.1
typing-extensions==3.10.0.0
Unidecode==1.1.1
urllib3==1.26.6
watchdog==2.1.6
wav2clip==0.1.0
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.37.1
yaspin==2.1.0
zipp==3.5.0
